requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
python-telegram-bot==20.8 